refactor: rename profiling files to have same casing as rest
