+++ src/profiling/ProfilingContextManager.ts
