+++ src/profiling/index.ts
@@ -25,9 +25,9 @@ import {
   ProfilingOptions,
   ProfilingStartOptions,
 } from './types';
-import { ProfilingContextManager } from './profiling_contextmanager';
-import { OTLPProfilingExporter } from './otlp_exporter';
-import { DebugExporter } from './debug_exporter';
+import { ProfilingContextManager } from './ProfilingContextManager';
+import { OTLPProfilingExporter } from './OTLPProfilingExporter';
+import { DebugExporter } from './DebugExporter';
 
 export { ProfilingOptions };
